requirements.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==2.9.2
Django==3.1.2
django-choices==1.7.1
django-cors-headers==3.5.0
djangorestframework==3.12.1
jwcrypto==0.8
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
